+++ ali/common/src/main/java/com/yanny/ali/plugin/common/trades/TradeNode.java
@@ -13,13 +13,14 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.Comparator;
 import java.util.List;
+import java.util.function.IntUnaryOperator;
 
 public class TradeNode extends ListNode {
     public static final ResourceLocation ID = Utils.modLoc("trade");
 
     private final TooltipNode tooltip;
 
-    public TradeNode(IServerUtils utils, Int2ObjectMap<VillagerTrades.ItemListing[]> itemListingMap, boolean isWanderingTrader) {
+    public TradeNode(IServerUtils utils, Int2ObjectMap<VillagerTrades.ItemListing[]> itemListingMap, IntUnaryOperator offersPerLevel) {
         List<Int2ObjectMap.Entry<VillagerTrades.ItemListing[]>> entries = itemListingMap.int2ObjectEntrySet()
                 .stream()
                 .sorted(Comparator.comparingInt(Int2ObjectMap.Entry::getIntKey))
@@ -27,7 +28,7 @@ public class TradeNode extends ListNode {
 
         for (Int2ObjectMap.Entry<VillagerTrades.ItemListing[]> entry : entries) {
             if (entry.getValue().length > 0) {
-                addChildren(new TradeLevelNode(utils, entry.getIntKey(), entry.getValue(), isWanderingTrader));
+                addChildren(new TradeLevelNode(utils, entry.getIntKey(), entry.getValue(), offersPerLevel.applyAsInt(entry.getIntKey())));
             }
         }